Core/Battleground: Fix DB load.

+++ src/server/game/Entities/Player/Player.cpp
@@ -19165,7 +19165,7 @@ void Player::_LoadBGData(PreparedQueryResult result)
     m_bgData.taxiPath[0]  = fields[7].GetUInt32();
     m_bgData.taxiPath[1]  = fields[8].GetUInt32();
     m_bgData.mountSpell   = fields[9].GetUInt32();
-    m_bgData.m_LastActiveSpec = fields[10].GetUInt8();
+    m_bgData.m_LastActiveSpec = fields[10].GetUInt16();
 }
 
 bool Player::LoadPositionFromDB(uint32& mapid, float& x, float& y, float& z, float& o, bool& in_flight, uint64 guid)
